requirements.txt
sympy==1.11.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
plotly==5.14.1
streamlit-extras==0.2.7
streamlit-code-editor==0.1.8
hydralit-components==1.0.10
streamlit-authenticator==0.2.2
streamlit-calendar==0.5.0
streamlit-card==0.0.5
streamlit-elements==0.1.0
streamlit-wordcloud==0.1.0
streamlit-antd-components==0.2.1
streamlit-on-Hover-tabs==1.0.1
streamlit-pandas-profiling==0.1.3
mitosheet==0.1.524
xata
streamlit-keplergl
extra-streamlit-components
